+++ frontend/app/app/pages/index.vue
@@ -28,7 +28,7 @@ import CommandBar from '~/components/layout/CommandBar.vue'
 import IntegrationsHub from '~/components/layout/IntegrationsHub.vue'
 import WorkspaceSettingsPanel from '~/components/settings/WorkspaceSettingsPanel.vue'
 import ObservabilityConnectionPanel from '~/components/settings/ObservabilityConnectionPanel.vue'
-import ModelDefaultsPanel from '~/components/settings/ModelDefaultsPanel.vue'
+import ModelConfigurationPanel from '~/components/settings/ModelConfigurationPanel.vue'
 import LocalModelEndpointsPanel from '~/components/settings/LocalModelEndpointsPanel.vue'
 import OpenRouterCatalogPanel from '~/components/settings/OpenRouterCatalogPanel.vue'
 import VendorCredentialsModal from '~/components/providers/VendorCredentialsModal.vue'
@@ -119,7 +119,7 @@ watch(
       <IntegrationsHub />
       <WorkspaceSettingsPanel />
       <ObservabilityConnectionPanel />
-      <ModelDefaultsPanel />
+      <ModelConfigurationPanel />
       <LocalModelEndpointsPanel />
       <OpenRouterCatalogPanel />
       <VendorCredentialsModal />